app/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector>=0.2.2
sentence-transformers
torch
sentencepiece>=0.1.99
protobuf>=3.20.0